+++ src/brasil/gov/tiles/tiles/audiogallery.py
@@ -1,8 +1,8 @@
 # -*- coding: utf-8 -*-
 from future.builtins import range  # isort:skip
 from brasil.gov.tiles import _ as _
-from brasil.gov.tiles.tiles.list import IListTile
-from brasil.gov.tiles.tiles.list import ListTile
+from collective.cover.tiles.list import IListTile
+from collective.cover.tiles.list import ListTile
 from plone import api
 from plone.tiles.interfaces import ITileDataManager
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile